Change matrix impl from 2d array to table
believe teh same way, except that the table will only
contain releveant information while the 2d array
will store a lot of useless infomration. Most node
are not adjcent to node x, why should we hold data
of all the nodes NOT adjacent to x when we can simply
jsut store the data that ARE adjacent to x.