Optimize the parsing for OO, modify the Node class
the main problem before was i was looping over a list of
nodes to look for an index. this step is repated alot!
so now we just put all the indexes inside a map, and thus
the lookup is MUCH faster.

Add some utility methods and fields to the Node class for the
OO approach